+++ ModelX/Measure/Temperature.cs
@@ -7,42 +7,42 @@ namespace ModelX.Measure
     public class Temperature : IMeasure
     {
         public Temperature() { }
-        public Temperature(double value) : this(value, Unit.Temperature.Celsius) { }
+        public Temperature(decimal value) : this(value, Unit.Temperature.Celsius) { }
     
-        public Temperature(double value, Enum unit)
+        public Temperature(decimal value, Enum unit)
         {
             Celsius = unit switch
             {
                 Unit.Temperature.Celsius    => value,
-                Unit.Temperature.Kelvin     => value - 273.15d,
-                Unit.Temperature.Fahrenheit => (value - 32d) * 5 / 9,
-                Unit.Temperature.Rankine    => (value - 273.15d) * 5 / 9,
-                Unit.Temperature.Newton     => value / 0.33d,
-                Unit.Temperature.Romer      => (value - 7.5d) * 40 / 21,
-                Unit.Temperature.Reaumur    => value * 5d / 4,
+                Unit.Temperature.Kelvin     => value - 273.15m,
+                Unit.Temperature.Fahrenheit => (value - 32) * 5 / 9,
+                Unit.Temperature.Rankine    => (value - 273.15m) * 5 / 9,
+                Unit.Temperature.Newton     => value / 0.33m,
+                Unit.Temperature.Romer      => (value - 7.5m) * 40m / 21m,
+                Unit.Temperature.Reaumur    => value * 5 / 4,
                 Unit.Temperature.Delisle    => 100 - value * 2 / 3,
                 _ => throw new NotSupportedException()
             };
         }
 
         [JsonProperty]
-        public double Celsius { get; set; } = 0d;
+        public decimal Celsius { get; set; } = 0;
         [JsonProperty]
-        public double Kelvin { get => Celsius + 273.15d; }
+        public decimal Kelvin { get => Celsius + 273.15m; }
         [JsonProperty]
-        public double Fahrenheit { get => Celsius * (9d / 5) + 32d; }
+        public decimal Fahrenheit { get => Celsius * (9 / 5) + 32; }
         [JsonProperty]
-        public double Rankine { get => Kelvin * (9d / 5); }
+        public decimal Rankine { get => Kelvin * (9m / 5m); }
         [JsonProperty]
-        public double Newton { get => 0.33d * Celsius; }
+        public decimal Newton { get => 0.33m * Celsius; }
         [JsonProperty]
-        public double Romer { get => Celsius * 21d / 40 + 7.5; }
+        public decimal Romer { get => Celsius * 21m / 40m + 7.5m; }
         [JsonProperty]
-        public double Reaumur { get => 0.8d * Celsius; }
+        public decimal Reaumur { get => 0.8m * Celsius; }
         [JsonProperty]
-        public double Delisle { get => (100 - Celsius) * 3d / 2; }
+        public decimal Delisle { get => (100 - Celsius) * 3m / 2; }
 
-        public double Result<T>(T unit) where T : Enum
+        public decimal Result<T>(T unit) where T : Enum
         {
             return unit switch
             {
